cleanup

+++ build/08-clean.ps1
@@ -1,10 +1,6 @@
-param (
-	[Parameter(Mandatory=$true)]
-	[string]
-	$Root
-)
-
 $Out = "$Root\out"
 Remove-Item "$Out\artifacts\signingbundle.zip" -Force
-Remove-Item "$Out\plugins.xml" -Force
+if ($BuildPlugins) {
+	Remove-Item "$Out\plugins.xml" -Force
+}
 Remove-Item "$Out\temp" -Force
